Separate stats filter utilities to reuse for page pruning
